requirements.txt
Flask
tensorflow 
pillow
gunicorn
keras